moving GetRandomNumber to a unique place
